test/pyproject.toml
[project]
name = "paic-python-starter"
version = "0.1.1"
description = "PAIC Python Starter"
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "anthropic>=0.39.0",
    "google-generativeai>=0.8.3",
    "llm>=0.18",
    "llm-claude-3>=0.9",
    "llm-gemini>=0.4.2",
    "ollama>=0.4.1",
    "openai[realtime]>=1.59.0",
    "mistralai>=0.2.3",
    "groq>=0.7.0",
    "pydantic>=2.10.2",
    "pytest>=8.3.3",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "typer>=0.13.1",
    "rich>=13.9.4",
    "realtimestt>=0.3.93",
    "dpath>=2.2.0",
    "elevenlabs>=1.50.3",
    "realtimetts[system]==0.4.40",
    "pyttsx3>=2.98",
    "gradio>=4.31.0",
    "scapy>=2.5.0",
    "psutil>=5.9.8",
    "dnspython>=2.6.1",
    "requests>=2.31.0",
    "playwright>=1.44.0",
    "browser-use>=0.1.41",
    "langchain-ollama>=0.0.11",
    "pillow>=10.3.0",
    "pvporcupine>=3.0.1",
    "sounddevice>=0.4.6",
    "python-nmap>=0.7.2",
    "shodan>=1.29.0",
    "censys-search>=2.2.5",
    "celery[redis]>=5.3.6",
    "redis>=5.0.0",
]
